requirements.txt
plaid-python~=15.0.0
SQLAlchemy~=2.0.18
requests~=2.31.0
schedule~=1.2.0